Restore x64 classic editor tools

+++ src/game/server/application/SwgDraftSchematicEditor/src/win32/DialogConsole.cpp
@@ -468,7 +468,7 @@ void DialogConsole::WriteToPipe( LPCTSTR line )
 {
 	DWORD dwWritten;
 
-	WriteFile( hChildStdinWrDup, line, _tcslen(line)*sizeof(TCHAR), 
+	WriteFile( hChildStdinWrDup, line, static_cast<DWORD>(_tcslen(line) * sizeof(TCHAR)),
 			&dwWritten, NULL );
 }
 
@@ -503,7 +503,7 @@ void DialogConsole::AddTexts(LPCTSTR string)
 
 void DialogConsole::MoveToEnd()
 {
-	int nLen = GetShellEditCtrl().SendMessage( WM_GETTEXTLENGTH );
+	int nLen = static_cast<int>(GetShellEditCtrl().SendMessage( WM_GETTEXTLENGTH ));
 	GetShellEditCtrl().SetSel( nLen,nLen );
 }
